+++ src/charts/PieChart/PieChart_snapshot.spec.js
@@ -1,14 +1,17 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import PieChart from './PieChart';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('PieChart Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
     
-      const wrapper1 = shallow(
+      const wrapper1 = mount(
         <PieChart
           title="Distribution of themes"
           data={require('../fixtures/pie-chart.js').data}
@@ -17,7 +20,7 @@ describe('PieChart Component', () => {
       );
       expect(wrapper1).toMatchSnapshot();
 
-      const wrapper2 = shallow(
+      const wrapper2 = mount(
         <PieChart
           title="Distribution of themes"
           data={require('../fixtures/pie-chart-custom-colors.js').data}
@@ -26,7 +29,7 @@ describe('PieChart Component', () => {
       );
       expect(wrapper2).toMatchSnapshot();
 
-      const wrapper3 = shallow(
+      const wrapper3 = mount(
         <PieChart
           data={require('../fixtures/pie-chart-one.js').data}
           colors={require('../fixtures/pie-chart-one.js').colors}