+++ src/components/ColorSelector/ColorSelector_snapshot.spec.js
@@ -1,14 +1,17 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import ColorSelector from './ColorSelector';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('ColorSelector Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
     
-      const wrapper1 = shallow(
+      const wrapper1 = mount(
         <ColorSelector
           color="#e25454"
           disableInteraction={true}
@@ -16,7 +19,7 @@ describe('ColorSelector Component', () => {
       );
       expect(wrapper1).toMatchSnapshot();
 
-      const wrapper2 = shallow(
+      const wrapper2 = mount(
         <ColorSelector
           color="#5F9DC7"
           onClick={() => alert('Clicked!')}
@@ -24,14 +27,14 @@ describe('ColorSelector Component', () => {
       );
       expect(wrapper2).toMatchSnapshot();
 
-      const wrapper3 = shallow(
+      const wrapper3 = mount(
         <ColorSelector
           patternImage="images/patternImage.jpg"
         />
       );
       expect(wrapper3).toMatchSnapshot();
 
-      const wrapper4 = shallow(
+      const wrapper4 = mount(
         <ColorSelector
           color="#aecc76"
           text="Lime green"
@@ -39,7 +42,7 @@ describe('ColorSelector Component', () => {
       );
       expect(wrapper4).toMatchSnapshot();
 
-      const wrapper5 = shallow(
+      const wrapper5 = mount(
         <ColorSelector
           color="#aecc76"
           selected={true}
@@ -48,7 +51,7 @@ describe('ColorSelector Component', () => {
       );
       expect(wrapper5).toMatchSnapshot();
 
-      const wrapper6 = shallow(
+      const wrapper6 = mount(
         <ColorSelector
           color="#e25454"
           fixedSize='40px'
@@ -57,7 +60,7 @@ describe('ColorSelector Component', () => {
       );
       expect(wrapper6).toMatchSnapshot();
 
-      const wrapper7 = shallow(
+      const wrapper7 = mount(
         <ColorSelector />
       );
       expect(wrapper7).toMatchSnapshot();