+++ src/components/Dropdown/Dropdown_snapshot.spec.js
@@ -1,13 +1,17 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import Dropdown from './Dropdown';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('Dropdown Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
-      const wrapper1 = shallow(
+    
+      const wrapper1 = mount(
         <Dropdown label="Don't push me">
           <Dropdown.Item>
             <nobr>...cause I'm close to the edge</nobr>
@@ -28,7 +32,7 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper1).toMatchSnapshot();
 
-      const wrapper2 = shallow(
+      const wrapper2 = mount(
         <Dropdown label="Don't push me" size="large">
           <Dropdown.Item>
             <nobr>...cause I'm close to the edge</nobr>
@@ -49,7 +53,7 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper2).toMatchSnapshot();
 
-      const wrapper3 = shallow(
+      const wrapper3 = mount(
         <Dropdown label="Don't push me" persist>
           <Dropdown.Item onSelect={() => alert('Clicked 1')}>
             <nobr>...cause I'm close to the edge</nobr>
@@ -70,7 +74,7 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper3).toMatchSnapshot();
 
-      const wrapper4 = shallow(
+      const wrapper4 = mount(
         <Dropdown label="Don't push me" preventOutOfBounds>
           <Dropdown.Item>
             <nobr>...cause I'm close to the edge</nobr>
@@ -91,7 +95,7 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper4).toMatchSnapshot();
 
-      const wrapper5 = shallow(
+      const wrapper5 = mount(
         <Dropdown label="Don't push me" preventOutOfBounds={false}>
           <Dropdown.Item>
             <nobr>...cause I'm close to the edge</nobr>
@@ -112,14 +116,14 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper5).toMatchSnapshot();
 
-      const wrapper6 = shallow(
+      const wrapper6 = mount(
         <Dropdown label="Push Me" persist>
           <nobr>You can place any content you prefer here.</nobr>
         </Dropdown>
       );
       expect(wrapper6).toMatchSnapshot();
 
-      const wrapper7 = shallow(
+      const wrapper7 = mount(
         <Dropdown invalid label="Don't push me">
           <Dropdown.Item>
             <nobr>...cause I'm close to the edge</nobr>
@@ -128,35 +132,35 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper7).toMatchSnapshot();
 
-      const wrapper8 = shallow(
+      const wrapper8 = mount(
         <Dropdown selected={true} label="Push Me">
           <Dropdown.Item>Check that checkmark</Dropdown.Item>
         </Dropdown>
       );
       expect(wrapper8).toMatchSnapshot();
 
-      const wrapper9 = shallow(
+      const wrapper9 = mount(
         <Dropdown loading={true} label="Something is happening">
           <Dropdown.Item>Hold on I'm loading</Dropdown.Item>
         </Dropdown>
       );
       expect(wrapper9).toMatchSnapshot();
 
-      const wrapper10 = shallow(
+      const wrapper10 = mount(
         <Dropdown disabled={true}>
           <Dropdown.Item>I'll never be seen</Dropdown.Item>
         </Dropdown>
       );
       expect(wrapper10).toMatchSnapshot();
 
-      const wrapper11 = shallow(
+      const wrapper11 = mount(
         <Dropdown label={<h3>h3 label</h3>}>
           <Dropdown.Item>Hey there</Dropdown.Item>
         </Dropdown>
       );
       expect(wrapper11).toMatchSnapshot();
 
-      const wrapper12 = shallow(
+      const wrapper12 = mount(
         <Dropdown open={true}>
           <Dropdown.Item>
             <nobr>You can't close me</nobr>
@@ -165,7 +169,7 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper12).toMatchSnapshot();
 
-      const wrapper13 = shallow(
+      const wrapper13 = mount(
         <Dropdown open={false} label="You can't open me">
           <Dropdown.Item>
             <nobr>You will never see me</nobr>
@@ -174,7 +178,7 @@ describe('Dropdown Component', () => {
       );
       expect(wrapper13).toMatchSnapshot();
 
-      const wrapper14 = shallow(
+      const wrapper14 = mount(
         <Dropdown label="Don't push me" fluid={true}>
           <Dropdown.Item onSelect={() => alert('Clicked 1')}>
             <nobr>...cause I'm close to the edge</nobr>
@@ -194,6 +198,7 @@ describe('Dropdown Component', () => {
         </Dropdown>
       );
       expect(wrapper14).toMatchSnapshot();
+
     });
   });
 });