+++ src/components/OverflowMenu/OverflowMenu_snapshot.spec.js
@@ -1,13 +1,17 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import OverflowMenu from './OverflowMenu';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('OverflowMenu Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
-      const wrapper1 = shallow(
+    
+      const wrapper1 = mount(
         <OverflowMenu>
           <OverflowMenu.Item onClick={() => alert('this.edit()')}>Edit</OverflowMenu.Item>
           <OverflowMenu.Item onClick={() => alert('this.delete()')}>Delete</OverflowMenu.Item>
@@ -15,7 +19,7 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper1).toMatchSnapshot();
 
-      const wrapper2 = shallow(
+      const wrapper2 = mount(
         <OverflowMenu size="medium">
           <OverflowMenu.Item onClick={() => alert('this.edit()')}>Edit</OverflowMenu.Item>
           <OverflowMenu.Item onClick={() => alert('this.delete()')}>Delete</OverflowMenu.Item>
@@ -23,7 +27,7 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper2).toMatchSnapshot();
 
-      const wrapper3 = shallow(
+      const wrapper3 = mount(
         <OverflowMenu size="large">
           <OverflowMenu.Item onClick={() => alert('this.edit()')}>Edit</OverflowMenu.Item>
           <OverflowMenu.Item onClick={() => alert('this.delete()')}>Delete</OverflowMenu.Item>
@@ -31,14 +35,14 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper3).toMatchSnapshot();
 
-      const wrapper4 = shallow(
+      const wrapper4 = mount(
         <OverflowMenu persist>
           <p>Some sample text</p>
         </OverflowMenu>
       );
       expect(wrapper4).toMatchSnapshot();
 
-      const wrapper5 = shallow(
+      const wrapper5 = mount(
         <OverflowMenu position="left">
           <OverflowMenu.Item>Edit</OverflowMenu.Item>
           <OverflowMenu.Item>Delete</OverflowMenu.Item>
@@ -46,7 +50,7 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper5).toMatchSnapshot();
 
-      const wrapper6 = shallow(
+      const wrapper6 = mount(
         <OverflowMenu appearance="primary">
           <OverflowMenu.Item>Edit</OverflowMenu.Item>
           <OverflowMenu.Item>Delete</OverflowMenu.Item>
@@ -54,7 +58,7 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper6).toMatchSnapshot();
 
-      const wrapper7 = shallow(
+      const wrapper7 = mount(
         <OverflowMenu appearance="danger">
           <OverflowMenu.Item>Edit</OverflowMenu.Item>
           <OverflowMenu.Item>Delete</OverflowMenu.Item>
@@ -62,7 +66,7 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper7).toMatchSnapshot();
 
-      const wrapper8 = shallow(
+      const wrapper8 = mount(
         <OverflowMenu color="red">
           <OverflowMenu.Item>Action 1</OverflowMenu.Item>
           <OverflowMenu.Item>Action 2</OverflowMenu.Item>
@@ -70,7 +74,7 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper8).toMatchSnapshot();
 
-      const wrapper9 = shallow(
+      const wrapper9 = mount(
         <OverflowMenu activeColor="blue">
           <OverflowMenu.Item>Action 1</OverflowMenu.Item>
           <OverflowMenu.Item>Action 2</OverflowMenu.Item>
@@ -78,7 +82,7 @@ describe('OverflowMenu Component', () => {
       );
       expect(wrapper9).toMatchSnapshot();
 
-      const wrapper10 = shallow(
+      const wrapper10 = mount(
         <OverflowMenu open={true}>
           <OverflowMenu.Title>You can't close me</OverflowMenu.Title>
           <OverflowMenu.Item>Action 1</OverflowMenu.Item>
@@ -86,6 +90,7 @@ describe('OverflowMenu Component', () => {
         </OverflowMenu>
       );
       expect(wrapper10).toMatchSnapshot();
+
     });
   });
 });