+++ src/components/RadioButton/RadioButton_snapshot.spec.js
@@ -1,23 +1,36 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import RadioButton from './RadioButton';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('RadioButton Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
-      const wrapper1 = shallow(<RadioButton label="Radio label" />);
+    
+      const wrapper1 = mount(
+        <RadioButton label="Radio label" />
+      );
       expect(wrapper1).toMatchSnapshot();
 
-      const wrapper2 = shallow(<RadioButton selected label="Radio label" />);
+      const wrapper2 = mount(
+        <RadioButton selected label="Radio label" readOnly />
+      );
       expect(wrapper2).toMatchSnapshot();
 
-      const wrapper3 = shallow(<RadioButton selected={false} label="Radio label" />);
+      const wrapper3 = mount(
+        <RadioButton selected={false} label="Radio label" />
+      );
       expect(wrapper3).toMatchSnapshot();
 
-      const wrapper4 = shallow(<RadioButton disabled label="Radio label" />);
+      const wrapper4 = mount(
+        <RadioButton disabled label="Radio label" />
+      );
       expect(wrapper4).toMatchSnapshot();
+
     });
   });
 });