+++ src/components/Text/Text_snapshot.spec.js
@@ -1,18 +1,21 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import Text from './Text';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('Text Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
-      const wrapper1 = shallow(
-        <Text bold size="XXXL" renderAs="h1">
-          Some text in the h1 style
-        </Text>
+    
+      const wrapper1 = mount(
+        <Text bold size="XXXL" renderAs="h1">Some text in the h1 style</Text>
       );
       expect(wrapper1).toMatchSnapshot();
+
     });
   });
 });