+++ src/globals/chart-styles.js
@@ -1,12 +1,11 @@
-import { injectGlobal } from 'styled-components';
+import { createGlobalStyle } from 'styled-components';
 import { theme } from '../theme';
 
-const injectGlobalStyles = () =>
-  injectGlobal`
+const ChartGlobalStyle = createGlobalStyle`
     @import 'https://code.highcharts.com/css/highcharts.css';
     .highcharts-container {
           font-family: ${theme.fontFamily};
     }
   `;
 
-injectGlobalStyles();
+export default ChartGlobalStyle;