+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## [Unreleased]
 
+### Changed
+
+- Add a missing `forwardRef` on [Dialog](/#/Components?id=dialog).
+- Improve test setup and snapshots.
+
+### Breaking changes
+
+- Update to `styled-components@^4.2.0`, peer React version is changed to `16.3.0`.
+
 ## [5.7.2] - 2019-05-07
 
 ### Changed